+++ common/src/main/kotlin/com/cobblemon/mod/common/client/render/models/blockbench/pokemon/gen6/AmauraModel.kt
@@ -1,0 +1,57 @@
+/*
+ * Copyright (C) 2023 Cobblemon Contributors
+ *
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at https://mozilla.org/MPL/2.0/.
+ */
+
+package com.cobblemon.mod.common.client.render.models.blockbench.pokemon.gen6
+
+import com.cobblemon.mod.common.client.render.models.blockbench.frame.HeadedFrame
+import com.cobblemon.mod.common.client.render.models.blockbench.pokemon.PokemonPose
+import com.cobblemon.mod.common.client.render.models.blockbench.pokemon.PokemonPoseableModel
+import com.cobblemon.mod.common.entity.PoseType
+import net.minecraft.client.model.ModelPart
+import net.minecraft.util.math.Vec3d
+
+class AmauraModel (root: ModelPart) : PokemonPoseableModel(), HeadedFrame {
+    override val rootPart = root.registerChildWithAllChildren("amaura")
+    override val head = getPart("neck")
+
+    override val portraitScale = 2.0F
+    override val portraitTranslation = Vec3d(-0.25, -0.85, 0.0)
+
+    override val profileScale = 0.9F
+    override val profileTranslation = Vec3d(0.0, 0.37, 0.0)
+
+    lateinit var standing: PokemonPose
+    lateinit var walk: PokemonPose
+
+    override fun registerPoses() {
+        standing = registerPose(
+            poseName = "standing",
+            poseTypes = PoseType.STATIONARY_POSES + PoseType.UI_POSES,
+            transformTicks = 10,
+            idleAnimations = arrayOf(
+                singleBoneLook(),
+                bedrock("amaura", "ground_idle")
+            )
+        )
+
+        walk = registerPose(
+            poseName = "walk",
+            poseTypes = PoseType.MOVING_POSES,
+            transformTicks = 10,
+            idleAnimations = arrayOf(
+                singleBoneLook(),
+                bedrock("amaura", "ground_idle")
+            )
+        )
+    }
+
+//    override fun getFaintAnimation(
+//        pokemonEntity: PokemonEntity,
+//        state: PoseableEntityState<PokemonEntity>
+//    ) = if (state.isPosedIn(standing, walk)) bedrockStateful("amaura", "faint") else null
+}