+++ common/src/main/kotlin/com/cobblemon/mod/common/client/render/models/blockbench/pokemon/gen8/OverqwilModel.kt
@@ -1,0 +1,80 @@
+/*
+ * Copyright (C) 2023 Cobblemon Contributors
+ *
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at https://mozilla.org/MPL/2.0/.
+ */
+
+package com.cobblemon.mod.common.client.render.models.blockbench.pokemon.gen8
+
+import com.cobblemon.mod.common.client.render.models.blockbench.animation.WaveAnimation
+import com.cobblemon.mod.common.client.render.models.blockbench.animation.WaveSegment
+import com.cobblemon.mod.common.client.render.models.blockbench.frame.HeadedFrame
+import com.cobblemon.mod.common.client.render.models.blockbench.pokemon.PokemonPose
+import com.cobblemon.mod.common.client.render.models.blockbench.pokemon.PokemonPoseableModel
+import com.cobblemon.mod.common.client.render.models.blockbench.pose.ModelPartTransformation
+import com.cobblemon.mod.common.client.render.models.blockbench.wavefunction.sineFunction
+import com.cobblemon.mod.common.entity.PoseType
+import net.minecraft.client.model.ModelPart
+import net.minecraft.util.math.Vec3d
+
+class OverqwilModel (root: ModelPart) : PokemonPoseableModel(), HeadedFrame {
+    override val rootPart = root.registerChildWithAllChildren("overqwil")
+    override val head = getPart("body")
+
+    val tail = getPart("tail")
+
+    val wtail = WaveSegment(tail, 7F)
+
+    override val portraitScale = 2.0F
+    override val portraitTranslation = Vec3d(-0.3, -1.2, 0.0)
+
+    override val profileScale = 0.8F
+    override val profileTranslation = Vec3d(0.0, 0.5, 0.0)
+
+    lateinit var standing: PokemonPose
+    lateinit var walk: PokemonPose
+
+    override fun registerPoses() {
+        val blink = quirk("blink") { bedrockStateful("overqwil", "blink").setPreventsIdle(false)}
+        standing = registerPose(
+            poseName = "standing",
+            poseTypes = PoseType.STATIONARY_POSES + PoseType.UI_POSES,
+            quirks = arrayOf(blink),
+            idleAnimations = arrayOf(
+                bedrock("overqwil", "ground_idle")
+            )
+        )
+
+        walk = registerPose(
+            poseName = "walk",
+            poseTypes = PoseType.MOVING_POSES,
+            quirks = arrayOf(blink),
+            idleAnimations = arrayOf(
+                bedrock("overqwil", "ground_idle"),
+                WaveAnimation(
+                    frame = this,
+                    waveFunction = sineFunction(
+                        period = 8F,
+                        amplitude = 0.4F
+                    ),
+                    basedOnLimbSwing = true,
+                    oscillationsScalar = 8F,
+                    head = head,
+                    rotationAxis = ModelPartTransformation.Y_AXIS,
+                    motionAxis = ModelPartTransformation.X_AXIS,
+                    headLength = 0.1F,
+                    segments = arrayOf(
+                        wtail
+                    )
+                )
+            )
+        )
+    }
+
+//    override fun getFaintAnimation(
+//        pokemonEntity: PokemonEntity,
+//        state: PoseableEntityState<PokemonEntity>
+//    ) = if (state.isPosedIn(standing, walk)) bedrockStateful("overqwil", "faint") else null
+}